requirements.txt
aniso8601==3.0.2
astroid==2.0.4
autopep8==1.4
click==6.7
colorama==0.3.9
Flask==1.0.2
Flask-JWT-Extended==3.13.0
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
